packege refactoring: cluster -> clients
