+++ vertical-pod-autoscaler/recommender/clients/cluster_feeder.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package cluster
+package clients
 
 import (
 	"github.com/golang/glog"
